refactor(adapters): split responses/mod.rs into focused sub-500-LOC modules (#17)
Split the 1961-LOC src/adapters/responses/mod.rs into 7 modules, each
under 500 LOC, to satisfy the file-size engineering standard:

- mod.rs (222) - adapter impl + forward dispatch
- http.rs (198, new) - HTTP transport + relay + SseParser
- pool.rs (425, new) - M10 Codex/ChatGPT account-pool failover
- request.rs (348, new) - request building, URL, identity headers, constants + tests
- websocket.rs (471, new) - WS turn machinery + tests
- ws_stream.rs (142, new) - WS event-stream relaying
- error.rs (274, new) - upstream/gateway error mapping + tests

Pure code-movement: 1368 code lines preserved verbatim, all 21 module
tests preserved and redistributed, clippy -D warnings clean, fmt
clean, 538 tests pass. Updated 5 contributor docs/ files to point
moved symbols at their new module files.

Closes #17
